Login page + general improvements

+++ front/src/services/exerciseService.js
@@ -1,16 +1,18 @@
 
 export const deleteExercise = (exerciseId) => {
-    return fetch('/api/exercises/' + exerciseId, {
-        method: 'DELETE',
-        headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json',
-            'Authorization': 'Bearer ' + localStorage.getItem('token')
-        },
-    })
-        .catch(error => {
-            console.error(error);
+    if (exerciseId) {
+        return fetch('/api/exercises/' + exerciseId, {
+            method: 'DELETE',
+            headers: {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json',
+                'Authorization': 'Bearer ' + localStorage.getItem('token')
+            },
         })
+            .catch(error => {
+                console.error(error);
+            })
+    }
 }
 
 export const getExercisesByTrainingId = (trainingId) => {
